+++ RELEASE_NOTES.md
@@ -1,6 +1,8 @@
 Release Notes
 =============
 ## 1.2.0-beta2
+* Container Groups: Support for image registry credentials for private registries.
+* Container Groups: Support for partial CPU cores.
 * Internal changes to better capture full resource IDs
 
 ## 1.2.0-beta1